refactor: extract DoorDash step-tracking from generic tool executor
